updating bad file
The November 2023 file needed to have the date updated to be in the correct Fair Shares week range.

I must have forgotten to collect a Jan 7, 2025 file so I made a copy of the Jan 14 file and am going to use that instead. Not ideal, but it will work for now.